fix route and create overflow tab shared component
